requirements.txt
# Copyright 2015 Solinea, Inc.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either expressed or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# This file is here because many Platforms as a Service look for
#        requirements.txt in the root directory of a project.
# requirements/base.txt is linked to it.

Django==1.6.11
South==1.0.2
arrow==0.4.4
celery[redis]==3.1.17
django-admin-bootstrapped==2.3.3
django-extensions==1.5.1
django-filter==0.9.2
django-polymorphic
django-rest-params==1.0.2
djangorestframework==3.1.0
djoser==0.2.1
drf-extensions==0.2.7
elasticsearch-curator
elasticsearch-dsl
elasticsearch
fabric==1.10.1
lockfile==0.9.1
networkx==1.9.1
pbr==0.10.8
psycopg2==2.5.4
python-cinderclient==1.1.1
python-glanceclient==0.17.0
python-keystoneclient==1.2.0  # v1.3.0 conflicts with cinderclient 1.1.1
python-neutronclient==2.3.11
python-novaclient==2.23.0
urllib3==1.8
wsgiref==0.1.2
